python3 frames2video.py $prefix $targetLanguage

+++ editFrames.py
@@ -22,21 +22,20 @@ def editFrame():
     files = list(filter(lambda file: prefix + "-" + str(count).zfill(3) + "-" in file, files))
     files = list(filter(lambda file: "-cm." not in file and "-cr." not in file and "-rs." not in file, files))
     files.sort()
-    #print(str(files))
-    #quit()
     for file in files:
         fileSplit = file.split(".")
         fileTokens = fileSplit[0].split("-")
         frameID = fileTokens[-1]
         backImagePrefix = "build/" + prefix + "-tr"
-        subprocess.call(["convert", backImagePrefix + ".jpg", "-crop", "800x483+230+0", backImagePrefix + "-cr.jpg"])
         imagePrefix = framePrefix + "-" + frameID
-        subprocess.call(["convert", imagePrefix + ".jpg", "-crop", "1660x1070+250+0", imagePrefix + "-cr.jpg"])
-        subprocess.call(["convert", imagePrefix + "-cr.jpg", "-resize", "37%", imagePrefix + "-rs.jpg"])
-        subprocess.call(["magick", "composite", "-gravity", "center", imagePrefix + "-rs.jpg", backImagePrefix + "-cr.jpg", imagePrefix + "-cm.jpg"])
-        os.system("rm -f " + imagePrefix + ".jpg")
-        os.system("rm -f " + imagePrefix + "-rs.jpg")
-        os.system("rm -f " + imagePrefix + "-cr.jpg")
+        if not path.exists(imagePrefix + "-cm.jpg"):
+            subprocess.call(["convert", backImagePrefix + ".jpg", "-crop", "800x482+230+0", backImagePrefix + "-cr.jpg"])
+            subprocess.call(["convert", imagePrefix + ".jpg", "-crop", "1660x1070+250+0", imagePrefix + "-cr.jpg"])
+            subprocess.call(["convert", imagePrefix + "-cr.jpg", "-resize", "37%", imagePrefix + "-rs.jpg"])
+            subprocess.call(["magick", "composite", "-gravity", "center", imagePrefix + "-rs.jpg", backImagePrefix + "-cr.jpg", imagePrefix + "-cm.jpg"])
+            os.system("rm -f " + imagePrefix + ".jpg")
+            os.system("rm -f " + imagePrefix + "-rs.jpg")
+            os.system("rm -f " + imagePrefix + "-cr.jpg")
  
 for line in lines:
     if "-->" in line:
